fix(mppi): the GPU planner's step index stopped being DevicePassable at rc2
`mppi_sample_actions_batched_kernel` took its horizon-step index as `Int`,
which no longer conforms to `DevicePassable` under Mojo 1.0.0rc2. The launch
site therefore failed to compile with a constraint error raised from inside
`SIMD`, which means `MPPIGPUBatched.plan_gpu` — and with it TD-MPC2's
`select_action_mpc` and EVERY `USE_MPC=True` example — has not built since the
rc2 bump.

Nothing in the package build instantiates that generic, so `pixi run build`
stayed green throughout. `tests/deep_agents/test_tdmpc2_agent_mpc_gpu.mojo` is
the gate that catches it, and it is not in CI; it passes again now.

Kernel param is `Int32` and the call site says `Int32(t)` — both are required,
since a bare `t` re-introduces the failure through implicit conversion.

+++ mojo_rl/planners/trajectory/mppi.mojo
@@ -847,7 +847,9 @@ def _run_mppi_iteration[
             std_tensor,
             act_step_tensor,
             all_actions_tensor,
-            t,
+            # Int32, not `t`: `Int` is not `DevicePassable` (rc2). The kernel
+            # param is Int32 and the CALL SITE has to say so.
+            Int32(t),
             Scalar[DType.uint32](step_seed),
             grid_dim=(MPPI_BLOCKS,),
             block_dim=(TPB,),
